=act Scheduler implementation separated from interface.
